+++ sdk/containerregistry/azure-containerregistry/azure/containerregistry/aio/_async_download_stream.py
@@ -6,6 +6,7 @@
 import hashlib
 from typing import AsyncIterator, AsyncContextManager, Awaitable, TypeVar, cast, Tuple, Dict, Any
 from typing_extensions import Protocol
+from azure.core.pipeline import PipelineResponse
 
 T = TypeVar('T', bound='AsyncDownloadBlobStream')
 
@@ -24,7 +25,7 @@ class AsyncDownloadBlobStream(
     def __init__(
         self,
         *,
-        response: AsyncIterator[bytes],
+        response: PipelineResponse,
         get_next: GetNext,
         blob_size: int,
         downloaded: int,
@@ -32,6 +33,7 @@ class AsyncDownloadBlobStream(
         chunk_size: int
     ) -> None:
         self._response = response
+        self._response_bytes = response.http_response.iter_bytes()
         self._next = get_next
         self._blob_size = blob_size
         self._downloaded = downloaded
@@ -49,15 +51,15 @@ class AsyncDownloadBlobStream(
         return self
 
     async def _yield_data(self) -> bytes:
-        data = await self._response.__anext__()
+        data = await self._response_bytes.__anext__()
         self._hasher.update(data)
         return data
 
-    async def _download_chunk(self) -> AsyncIterator[bytes]:
+    async def _download_chunk(self) -> PipelineResponse:
         end_range = self._downloaded + self._chunk_size
         range_header = f"bytes={self._downloaded}-{end_range}"
         next_chunk, headers = cast(
-            Tuple[AsyncIterator[bytes], Dict[str, str]],
+            Tuple[PipelineResponse, Dict[str, str]],
             await self._next(range_header=range_header)
         )
         self._downloaded += int(headers["Content-Length"])
@@ -66,11 +68,15 @@ class AsyncDownloadBlobStream(
     async def __anext__(self) -> bytes:
         try:
             return await self._yield_data()
-        except StopIteration:
+        except StopAsyncIteration:
             if self._downloaded >= self._blob_size:
                 computed_digest = "sha256:" + self._hasher.hexdigest()
                 if computed_digest != self._digest:
                     raise ValueError("The requested digest does not match the digest of the received blob.")
                 raise
             self._response = await self._download_chunk()
+            self._response_bytes = self._response.http_response.iter_bytes()
             return await self._yield_data()
+
+    async def close(self):
+        await self._response.http_response.close()